fix : <ktp_task_controller> error handling of graph sync response logics
